% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_narrative.R
\name{case_narrative}
\alias{case_narrative}
\title{Case narrative}
\usage{
case_narrative(filepath)
}
\arguments{
\item{filepath}{character string of file path}
}
\value{
data frame
}
\description{
Case narrative
}
