% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_case_narratives.R
\name{multiple_case_narratives}
\alias{multiple_case_narratives}
\title{Multiple case narratives}
\usage{
multiple_case_narratives(folderpath)
}
\arguments{
\item{folderpath}{character string of folder path containing only summary files}
}
\value{
data frame
}
\description{
Multiple case narratives
}
